modulizes calendar to be deportable in other pages
